Use readiness endpoint to determine CW readiness
https://github.com/eclipse/codewind/issues/65
